+++ DesignToolbox/DesignToolbox/Pages/Components/Button/ButtonConfiguration.swift
@@ -67,7 +67,7 @@ final class ButtonConfigurationModel: ComponentConfiguration {
     }
 
     private var coloredSurfaceCodeModifier: String {
-        onColoredSurface ? ".oudsColoredSurface(color: Color.orange)" : ""
+        onColoredSurface ? ".oudsColoredSurface(theme.colorModes.modeOnBrandPrimary)" : ""
     }
 
     override func updateCode() {
@@ -173,7 +173,7 @@ struct ButtonConfiguration: View {
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(model.style != .`default`)
 
-            Toggle("app_components_common_onColoredBackground_label", isOn: $model.onColoredSurface)
+            Toggle("app_components_common_onColoredSurface_label", isOn: $model.onColoredSurface)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)