+++ DesignToolbox/DesignToolbox/Pages/Components/Button/ButtonPage.swift
@@ -103,6 +103,6 @@ private struct ButtonDemo: View {
         }
         .disabled(!model.enabled)
         .padding(.all, theme.spaces.spaceFixedMedium)
-        .modifier(DesignToolboxColoredBackgroundModifier(coloredSurface: model.onColoredSurface))
+        .modifier(DesignToolboxColoredSurfaceModifier(coloredSurface: model.onColoredSurface))
     }
 }