+++ DesignToolbox/DesignToolbox/Pages/Utils/DesignToolboxColoredSurfaceModifier.swift
@@ -17,7 +17,7 @@ import SwiftUI
 /// Use the background modifier to place a view on a colored surface.
 /// Today, in the design kit application the color of the surface is based on the sementic token
 /// named: _color/surface/brand/primary_
-struct DesignToolboxColoredBackgroundModifier: ViewModifier {
+struct DesignToolboxColoredSurfaceModifier: ViewModifier {
 
     @Environment(\.theme) private var theme
 
@@ -25,7 +25,7 @@ struct DesignToolboxColoredBackgroundModifier: ViewModifier {
 
     func body(content: Content) -> some View {
         if coloredSurface {
-            content.oudsColoredSurface(colors: theme.colors.colorSurfaceBrandPrimary)
+            content.oudsColoredSurface(theme.colorModes.modeOnBrandPrimary)
         } else {
             content.oudsBackground(theme.colors.colorBgSecondary)
         }